200_application/requirements.txt
fastapi[standard]>=0.115.12
python-dotenv>=1.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-otlp>=1.20.0
azure-monitor-opentelemetry>=1.0.0
azure-ai-projects==1.0.0b7
azure-identity==1.21.0